"Add investment functionality for founders in ProfileAutre component"
This commit adds a new feature to the ProfileAutre component that allows founders to invest in a startup. The investment is done through an API call to '/api/store-flouci-account' and '/api/generate-payment' with the necessary data. A loading state is also added to indicate when the investment is in progress.